fix(legacy-button): rename test as recommended by @benoitf
